Standardize the clean task to delete the build.dir.
First step to defining all common tasks in a single build file.

This no longer calls the clean task on dependency projects, which is a
little annoying, but it shouldn't make much difference if the build task
for those projects depends on the clean task, which it probably should
(more standardization to come).